Got interrupts working!
Issue one was that my structure sizes were fucking things up, ex ((this_struct*)myStruct + 2) == &myStruct + (2*sizeof(this_struct)), which should have been obvious to begin with

Issue two was that the inline asm I had set up to load the IDT was incorrect, though fuck me if I know why. Point is, inspecting idtr using bochsdbg led to the realization that idtr was not getting set to 0x201030 which is what was really fucking everything up

Also, one minor fix was that I realized that some of my fuckuppery that was resulting ipossibly ito some of my early crashes but also definitively to such issues as screwed up output was that after moving the kernel to 1 meg my A20 check routine was suddenly flipping a bit in the middle of the kernel which was bodging things up real good. For now, the code is just disabled (because dosload is now doing the A20 shit anyhow) but I will repair it in the future.